requirements.txt
discord.py
python-dotenv
jinja2